Separate out presets and and some utils
