+++ demo/js/demo.js
@@ -1,10 +1,14 @@
 import * as GCodePreview from 'gcode-preview';
 import * as THREE from 'three';
+import { settingsPresets } from './presets.js';
+import { debounce, throttle, humanFileSize, storeSettings } from './utils.js';
+
+const defaultPreset = 'multicolor';
+const maxToolCount = 8;
 
 let gcodePreview;
 let favIcon;
 let thumb;
-const maxToolCount = 8;
 let toolCount = 4;
 let gcode;
 let renderProgressive = true;
@@ -44,106 +48,6 @@ const buildVolumeZ = document.getElementById('buildVolumeZ');
 const drawBuildVolume = document.getElementById('drawBuildVolume');
 const travelColor = document.getElementById('travel-color');
 const preferDarkMode = window.matchMedia('(prefers-color-scheme: dark)');
-
-const defaultPreset = 'multicolor';
-
-const settingsPresets = {
-  multicolor: {
-    file: 'gcodes/3DBenchy-Multi-part.gcode',
-    lineWidth: 1,
-    singleLayerMode: false,
-    renderExtrusion: true,
-    renderTubes: false,
-    extrusionColors: ['#CF439D', 'rgb(84,74,187)', 'white', 'rgb(83,209,104)'],
-    travel: false,
-    travelColor: '#00FF00',
-    highlightTopLayer: false,
-    topLayerColor: undefined,
-    lastSegmentColor: undefined,
-    drawBuildVolume: true,
-    buildVolume: {
-      x: 180,
-      y: 180,
-      z: 200
-    }
-  },
-  mach3: {
-    file: 'gcodes/mach3.gcode',
-    lineWidth: 1,
-    singleLayerMode: false,
-    renderExtrusion: false,
-    renderTubes: false,
-    extrusionColors: [],
-    travel: true,
-    travelColor: '#00FF00',
-    highlightTopLayer: false,
-    topLayerColor: undefined,
-    lastSegmentColor: undefined,
-    drawBuildVolume: true,
-    buildVolume: {
-      x: 20,
-      y: 20,
-      z: ''
-    }
-  },
-  arcs: {
-    file: 'gcodes/screw.gcode',
-    lineWidth: 2,
-    singleLayerMode: true,
-    renderExtrusion: true,
-    renderTubes: true,
-    extrusionColors: ['rgb(83,209,104)'],
-    travel: false,
-    travelColor: '#00FF00',
-    highlightTopLayer: false,
-    topLayerColor: undefined,
-    lastSegmentColor: undefined,
-    drawBuildVolume: true,
-    buildVolume: {
-      x: 200,
-      y: 200,
-      z: 180
-    }
-  },
-  'vase-mode': {
-    file: 'gcodes/vase.gcode',
-    lineWidth: 1,
-    singleLayerMode: true,
-    renderExtrusion: true,
-    renderTubes: true,
-    extrusionColors: ['rgb(84,74,187)'],
-    travel: false,
-    travelColor: '#00FF00',
-    highlightTopLayer: true,
-    topLayerColor: '#40BFBF',
-    lastSegmentColor: '#ffffff',
-    drawBuildVolume: true,
-    buildVolume: {
-      x: 200,
-      y: 200,
-      z: 220
-    }
-  },
-  'travel-moves': {
-    file: 'gcodes/plant-sign.gcode',
-    lineWidth: 2,
-    singleLayerMode: false,
-    renderExtrusion: true,
-    renderTubes: true,
-    extrusionColors: ['#777777'],
-    travel: true,
-    travelColor: '#00FF00',
-    highlightTopLayer: true,
-    topLayerColor: '#aaaaaa',
-    lastSegmentColor: undefined,
-    drawBuildVolume: true,
-    buildVolume: {
-      x: 200,
-      y: 200,
-      z: 220
-    }
-  }
-};
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars, no-unused-vars
 export function initDemo() {
@@ -326,7 +230,7 @@ export function initDemo() {
 
     preview.render();
 
-    storeSettings();
+    storeSettings(gcodePreview);
   }
 
   buildVolumeX.addEventListener('input', updateBuildVolume);
@@ -483,19 +387,6 @@ export function initDemo() {
   return preview;
 }
 
-function storeSettings() {
-  localStorage.setItem(
-    'settings',
-    JSON.stringify({
-      buildVolume: {
-        x: gcodePreview.buildVolume.x,
-        y: gcodePreview.buildVolume.y,
-        z: gcodePreview.buildVolume.z
-      }
-    })
-  );
-}
-
 function updateUI() {
   startLayer.setAttribute('max', gcodePreview.layers.length);
   endLayer.setAttribute('max', gcodePreview.layers.length);
@@ -560,11 +451,6 @@ async function startLoadingProgressive(gcode) {
   endLayer.removeAttribute('disabled');
 }
 
-function humanFileSize(size) {
-  var i = Math.floor(Math.log(size) / Math.log(1024));
-  return (size / Math.pow(1024, i)).toFixed(2) * 1 + ' ' + ['B', 'kB', 'MB', 'GB', 'TB'][i];
-}
-
 function setFavicons(favImg) {
   const headTitle = document.querySelector('head');
   const setFavicon = document.createElement('link');
@@ -572,23 +458,6 @@ function setFavicons(favImg) {
   setFavicon.setAttribute('href', favImg);
   headTitle.appendChild(setFavicon);
 }
-
-let throttleTimer;
-const throttle = (callback, time) => {
-  if (throttleTimer) return;
-  throttleTimer = true;
-  setTimeout(() => {
-    callback();
-    throttleTimer = false;
-  }, time);
-};
-
-// debounce function
-let debounceTimer;
-const debounce = (callback) => {
-  if (debounceTimer) clearTimeout(debounceTimer);
-  debounceTimer = setTimeout(callback, 300);
-};
 
 function showExtrusionColors() {
   // loop through inputs and show/hide them
